First rough version of sane mail forwarding
With a helper script, mutt forwards mails like for e.g. thunderbird.
I.e., the plain text mail is sent inline and can be edited, all
attachments are forwarded as attachments of the new mail.

Works so far, but adds an unnecessary layer of mime parts and there
are no error checks. use with care.